# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 大陆农机配件公共 REST API
  version: 1.0.0
  description: 新增移动端与供求接口的语言无关契约；既有 /api 接口保持兼容。
servers:
  - url: /
paths:
...
  /api/admin/vendor-posts:
    get:
      operationId: listVendorPostsForRole
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      operationId: createVendorPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VendorPostInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
  /api/admin/vendor-posts/{id}/review:
    put:
      operationId: reviewVendorPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    cookieAuth: { type: apiKey, in: cookie, name: dlc_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
        data: {}
...
    VendorPostInput:
      type: object
      required: [postType, title, content]
      properties:
        postType: { type: string, enum: [update, case] }
        title: { type: string, maxLength: 160 }
        summary: { type: string, maxLength: 500 }
        content: { type: string }
        coverImage: { type: string }
        coverAssetId: { type: integer }